Logging test history after failure (#8)
Printing out the places where a thread switches context if a test
failures.

This task is not quite done, because not all line numbers are
instrumented yet - just field access. We should record all line numbers
a test visits so that we print the most accurate line number when there
is a context switch.