Instrumenting all class loaders to make class loading atomic
The previous approach to make class loading atomic by creating a
classloader and calling Thread.setContextClassLoader did not work,
because some loads will not use that class loader, for example loading
JDK classes or loading dependencies of classes that are already loaded
before the test.

Instead of this approach, we instrument *all* class loaders to mark the
loading as atomic.