Trying to track object assignments
As a first step to tracking which objects are confined to a thread,
trying to track what the owner and fieldValues are for every assignment.

This is currently failing, because apparently the constructors of
(non-static) inner classes assign the value of this$0 *before* calling
the super constructor. The JVM seems to get made if you try to pass
*this* (the instance of the inner class) to a method, saying
uninitializedThis cannot be assigned to java.lang.Object).

I'm not sure how to get around this. Skip tracking those assignments,
because they are final? What will that break?