Trying to ignore final fields
Ignoring final fields. This solves errors with anonymous class
constructors, which set the field synthetic field this$0 before the
super() call. The value of this cannot be based to another method before
the super() call, which means we can't instrument the operation of
setting this$0 on this.

Also, fixing the way we were handling static field sets.

This is still not working, we are seeing stack corruption issues with
the integration tests. Something is eating more stack than it should!