Setting the minimum version of java in generated class files
We are using a java 1.5 feature (LDC of a class constant) so we need to
set the java class file version to at least 1.5.

This fixes the previously observed issues with instrumenting static get
field calls when running through gradle (gradle adds an old version of
commons-lang to the classpath, which then could not be instrumented).