Adding a utility for finding reachable objects
This utility finds all objects reachable from a given object through non
static references.

However, this is using java 8 streams, which is making this harder to
debug, and the recursive nature of the stream is creating a potential
for stack overflow errors with large graphs. Checking this in to
reconsider the approach.