Avoiding recursion the reachable object finder
Using a queue of unvisited objects rather than recursive calls to avoid
stack overflows when traversing reachable objects.

Adding a filter to stop looking for reachable objects if we hit a shared
object. We are only interested in looking for all of the thread confined
objects reachable from a thread confined object.