Retransforming all loaded classes in agent
The FatesAgent was only retransforming a few specific classes. I
discovered that important classes like HashMap were already loaded, so
were never being instrumented.

However, the tests are now failing with a stack overflow error, because
the NonReentrantExecutionEventListener is dependent on some of these
classes before it can even do it's work. We need to improve that path to
not use core classes - maybe our own subpackaged ThreadLocal?