Not instrumenting ThreadLocal and WeakReference
In order to avoid reentering the instrumentation code while checking to
see if we are reentering the instrumentation code, we need to ignore
thread local and weak reference, because the
NonReentrantExecutionEventListener uses those classes.

Another approach would be to create our own thread local like class in
the core package, because nothing in the core package is instrumented.
That might be a cleaner approach.

Now the tests are failing with a hotspot error. However, it looks like
the issue is twofold
- There is some verification error of bytecode
- we have instrumented some errors like java.lang.LinkageError, which
  the JDK is trying to throw when it hits a verification error!