Working around max stack size corruption - #13
Asm was computing the maximum stack size for each method. However, for
some JDK methods - eg Class.getSimpleBinaryName, it adjusted the max
stack size *down*, which was incorrect.

I suspect this has something to do with missing or incorrect FRAME
information in the bytecode passed to the ClassFileTransformer - I don't
see that information in there. I'm not sure why.

Rather than use COMPUTE_MAXs, this change works around the code by just
increasing the max stack size to account for the maximum increase any of
our instrumentation might add, which is just a fixed 5 elements.

We can improve this later to only increase the stack size in methods
we've touched, or ideally only where we have increase the stack size in
the largest path.