Running tests in a separate process with agent (#2)
If ThreadFates.run is not executed within a JVM that has the
instrumentation agent registered, run the test in a separate process
with the instrumentation agent registered automically.

This allows the user to consumer the fates-threads jar just like any
other jar and use it in their tests like any other dependency, without
special build magic to add the instrumentation agent.

Refactoring the instrumentation code to be a separate jar. The agent now
has the minimal code to do bytecode instrumentation, the testing harness
plugs itself in at runtime later.