Cleaning up dependencies in publication
We were not generating the correct pom for the fates-threads project; it
was not listing the dependencies on the other projects.

The fates-instrumentation project has some interesting issues due to the
fact that we are creating a shadow jar with asm embedded (as the asm
project recommends). The configuration here seems to publish the
correctly shadowed jar and use that as a dependency in the
fates-threads.