Changing postValidation to wait for launched threads
Rather than failing the test if threads are still running, join them. If
the test initiated shutdown of the threads this will allow the test to
pass, if it did not we will detect a deadlock and fail the test anyway.
This ensures that all threads are cleaned up while being more lenient
about code that doesn't bother to wait for it's threads to finish.